+++ src/effects/mod.rs
@@ -1,1 +1,2 @@
+pub mod animate;
 pub mod blink;